+++ packages/chrome-extension/src/app/components/single/singleEditorView.tsx
